+++ src/route.rs
@@ -65,16 +65,22 @@ impl<S: 'static> Route<S> {
     /// Application::new()
     ///    .resource("/path", |r|
     ///       r.route()
-    ///          .p(pred::Get())
-    ///          .p(pred::Header("content-type", "text/plain"))
+    ///          .filter(pred::Get())
+    ///          .filter(pred::Header("content-type", "text/plain"))
     ///          .f(|req| HTTPOk)
     ///       )
     /// #      .finish();
     /// # }
     /// ```
-    pub fn p<T: Predicate<S> + 'static>(&mut self, p: T) -> &mut Self {
+    pub fn filter<T: Predicate<S> + 'static>(&mut self, p: T) -> &mut Self {
         self.preds.push(Box::new(p));
         self
+    }
+
+    #[doc(hidden)]
+    #[deprecated(since="0.4.1", note="please use `.filter()` instead")]
+    pub fn p<T: Predicate<S> + 'static>(&mut self, p: T) -> &mut Self {
+        self.filter(p)
     }
 
     /// Set handler object. Usually call to this method is last call